Created InstructionType.hpp, InstructionException.hpp and InstructionException.cpp files
