+++ include/Vandior/InstructionException.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include <stdexcept>
+#include "Token.hpp"
+
+namespace vnd {
+
+    class InstructionException : public std::runtime_error {
+    public:
+        InstructionException(const Token &token);
+    };
+
+}
